preserving state on LayerItem
