+++ static/backbone/js/api/views/maps/sidepanel/menus/layersMenu.js
@@ -1,16 +1,17 @@
 define(["marionette",
             "text!" + templateDir + "/sidepanel/menuItem.html",
         "underscore",
-        "jquery"
+        "jquery",
+        "collections/layers"
     ],
-    function (Marionette, projectItem, _, $) {
+    function (Marionette, projectItem, _, $, Layers) {
         'use strict';
         /**
          * Class that controls the available projects menu,
          * Extends Backbone.View.
-         * @class ProjectsMenu
+         * @class LayersMenu
          */
-        var ProjectsMenu = Marionette.CollectionView.extend({
+        var LayersMenu = Marionette.CollectionView.extend({
             /**
              * @lends localground.maps.views.ProjectsMenu#
              */
@@ -25,7 +26,7 @@ define(["marionette",
                 template: _.template(projectItem),
                 modelEvents: {'change': 'render'}
             }),
-            id: 'projects-menu',
+            id: 'layers-menu',
             /**
              * Initializes the project menu and fetches the available
              * projects from the Local Ground Data API.
@@ -38,31 +39,22 @@ define(["marionette",
             initialize: function (opts) {
                 //this.setElement(opts.el);
                 this.app = opts.app;
-                this.collection = opts.projects;
+                this.collection = new Layers();
                 this.childViewOptions.app = this.app;
-                this.listenTo(this.app.vent, 'toggle-project', this.toggleProject);
-                this.app.vent.trigger('load-projects', this.collection);
+                this.collection.fetch();
+                this.listenTo(this.app.vent, 'toggle-layer', this.toggleItem);
                 this.restoreState();
             },
 
-            onAddChild: function (childView) {
-                var project = childView.model;
-                if (this.state) {
-                    if (_.contains(this.state.activeProjects, project.get('id'))) {
-                        this.triggerToggleCheckbox({target: childView.el});
-                    }
-                }
-            },
             /**
              * Catches the div click event and ignores it
              * @param {Event} e
              */
             toggleCheckbox: function (e) {
-                var input = $(e.target).find('input').addBack().filter('input');
-                var checked = input.is(':checked'),
-                    projectId = input.val();
-                this.app.vent.trigger('toggle-project', projectId, checked);
-
+                var input = $(e.target).find('input').addBack().filter('input'),
+					checked = input.is(':checked'),
+                    id = input.val();
+                this.app.vent.trigger('toggle-layer', id, checked);
                 if (e.stopPropagation) {
                     e.stopPropagation();
                 }
@@ -75,11 +67,24 @@ define(["marionette",
                 }
             },
 
-            saveState: function () {
+            toggleItem: function (id, visible) {
+                var model = this.collection.get(id);
+                if (model) {
+                    model.set('isVisible', visible);
+                }
+                if (visible) {
+                    this.app.vent.trigger("add-layer", { layer: model });
+                } else {
+                    this.app.vent.trigger("remove-layer", { layer: model });
+                }
+                this.saveState();
+            },
+
+			saveState: function () {
                 this.app.saveState(
                     this.id,
                     {
-                        activeProjects: _.chain(this.collection.toJSON())
+                        activeLayers: _.chain(this.collection.toJSON())
                             .where({isVisible: true})
                             .pluck('id')
                             .value()
@@ -90,17 +95,7 @@ define(["marionette",
 
             restoreState: function () {
                 this.state = this.app.restoreState(this.id);
-            },
-
-            toggleProject: function (projectId, visible) {
-                var project = this.collection.get(projectId);
-                if (project) {
-                    project.set('isVisible', visible);
-                }
-                this.collection.trigger('selected-projects-change');
-                this.collection.trigger('toggleProject', projectId, visible);
-                this.saveState();
             }
         });
-        return ProjectsMenu;
+        return LayersMenu;
     });
