+++ static/backbone/js/api/views/maps/sidepanel/menus/projectsMenu.js
@@ -1,17 +1,16 @@
 define(["marionette",
             "text!" + templateDir + "/sidepanel/menuItem.html",
         "underscore",
-        "jquery",
-        "collections/layers"
+        "jquery"
     ],
-    function (Marionette, projectItem, _, $, Layers) {
+    function (Marionette, projectItem, _, $) {
         'use strict';
         /**
          * Class that controls the available projects menu,
          * Extends Backbone.View.
-         * @class LayersMenu
+         * @class ProjectsMenu
          */
-        var LayersMenu = Marionette.CollectionView.extend({
+        var ProjectsMenu = Marionette.CollectionView.extend({
             /**
              * @lends localground.maps.views.ProjectsMenu#
              */
@@ -26,6 +25,7 @@ define(["marionette",
                 template: _.template(projectItem),
                 modelEvents: {'change': 'render'}
             }),
+            id: 'projects-menu',
             /**
              * Initializes the project menu and fetches the available
              * projects from the Local Ground Data API.
@@ -38,22 +38,31 @@ define(["marionette",
             initialize: function (opts) {
                 //this.setElement(opts.el);
                 this.app = opts.app;
-                this.collection = new Layers();
+                this.collection = opts.projects;
                 this.childViewOptions.app = this.app;
-                this.collection.fetch();
-                this.listenTo(this.app.vent, 'toggle-layer', this.toggleLayer);
+                this.listenTo(this.app.vent, 'toggle-project', this.toggleItem);
+                this.app.vent.trigger('load-projects', this.collection);
                 this.restoreState();
             },
 
+            onAddChild: function (childView) {
+                var project = childView.model;
+                if (this.state) {
+                    if (_.contains(this.state.activeProjects, project.get('id'))) {
+                        this.triggerToggleCheckbox({target: childView.el});
+                    }
+                }
+            },
             /**
              * Catches the div click event and ignores it
              * @param {Event} e
              */
             toggleCheckbox: function (e) {
-                var input = $(e.target).find('input').addBack().filter('input'),
-					checked = input.is(':checked'),
-                    id = input.val();
-                this.app.vent.trigger('toggle-layer', id, checked);
+                var input = $(e.target).find('input').addBack().filter('input');
+                var checked = input.is(':checked'),
+                    projectId = input.val();
+                this.app.vent.trigger('toggle-project', projectId, checked);
+
                 if (e.stopPropagation) {
                     e.stopPropagation();
                 }
@@ -66,24 +75,11 @@ define(["marionette",
                 }
             },
 
-            toggleLayer: function (id, visible) {
-                var model = this.collection.get(id);
-                if (model) {
-                    model.set('isVisible', visible);
-                }
-                if (visible) {
-                    this.app.vent.trigger("add-layer", { layer: model });
-                } else {
-                    this.app.vent.trigger("remove-layer", { layer: model });
-                }
-                this.saveState();
-            },
-
-			saveState: function () {
+            saveState: function () {
                 this.app.saveState(
                     this.id,
                     {
-                        activeLayers: _.chain(this.collection.toJSON())
+                        activeProjects: _.chain(this.collection.toJSON())
                             .where({isVisible: true})
                             .pluck('id')
                             .value()
@@ -94,7 +90,17 @@ define(["marionette",
 
             restoreState: function () {
                 this.state = this.app.restoreState(this.id);
+            },
+
+            toggleItem: function (projectId, visible) {
+                var project = this.collection.get(projectId);
+                if (project) {
+                    project.set('isVisible', visible);
+                }
+                this.collection.trigger('selected-projects-change');
+                this.collection.trigger('toggleProject', projectId, visible);
+                this.saveState();
             }
         });
-        return LayersMenu;
+        return ProjectsMenu;
     });
