+++ static/backbone/js/templates/sidepanel/layerEntry.html
@@ -1,12 +1,13 @@
 <div class="layer-header">
-    <input class="check-all" type="checkbox">
+    <input class="check-all" type="checkbox" <% if(showOverlay) { %>checked<% } %>>
     <h4><%= name %></h4>
     <i class="fa fa-search fa-lg pull-right zoom-to-extent" title="zoom to extent"></i>
 </div>
 <% _.each(symbols, function(item){ %>
 
     <div class="item layer-item">
-        <input class="cb-layer-item" type="checkbox" value="<%= item.rule %>">
+        <input class="cb-layer-item" type="checkbox" value="<%= item.rule %>"
+            <% if(item.showOverlay) { %>checked<% } %>>
         <svg width="<%= item.width %>" viewBox="<%= item.viewBox %>"
             style="margin-top:3px;">
             <path fill="<%= item.color %>" d="<%= item.path %>"></path>
